+++ lib/barcode/barcode.dart
@@ -2,6 +2,7 @@ import "package:flutter/material.dart";
 
 import "package:flutter_speed_dial/flutter_speed_dial.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
+import "package:inventree/preferences.dart";
 import "package:one_context/one_context.dart";
 
 
@@ -11,6 +12,7 @@ import "package:inventree/helpers.dart";
 import "package:inventree/l10.dart";
 
 import "package:inventree/barcode/camera_controller.dart";
+import "package:inventree/barcode/wedge_controller.dart";
 import "package:inventree/barcode/controller.dart";
 import "package:inventree/barcode/handler.dart";
 import "package:inventree/barcode/tones.dart";
@@ -43,6 +45,19 @@ Future<Object?> scanBarcode(BuildContext context, {BarcodeHandler? handler}) asy
   handler ??= BarcodeScanHandler();
   
   InvenTreeBarcodeController controller = CameraBarcodeController(handler);
+
+  // Select barcode controller based on user preference
+  final int barcodeControllerType = await InvenTreeSettingsManager().getValue(INV_BARCODE_SCAN_TYPE, BARCODE_CONTROLLER_CAMERA) as int;
+
+  switch (barcodeControllerType) {
+    case BARCODE_CONTROLLER_WEDGE:
+      controller = WedgeBarcodeController(handler);
+      break;
+    case BARCODE_CONTROLLER_CAMERA:
+    default:
+      // Already set as default option
+      break;
+  }
 
   return Navigator.of(context).push(
     PageRouteBuilder(